split some large files into separate components
