+++ frontend/src/scenes/trends/ActionsGraph.js
@@ -1,166 +1,14 @@
 import React, { Component } from 'react'
-import LineGraph from './LineGraph';
 import api from '../../lib/api';
-import { toParams, fromParams, Loading, Card, CloseButton, selectStyle } from '../../lib/utils';
+import { toParams, fromParams, Card, CloseButton } from '../../lib/utils';
 import { Dropdown } from "../../lib/components/Dropdown";
-import { Link } from 'react-router-dom';
-import PropTypes from 'prop-types';
-import Select from 'react-select';
 import SaveToDashboard from '../../lib/components/SaveToDashboard';
 import PropertyFilters from '../../lib/components/PropertyFilter';
-import moment from 'moment';
 import DateFilter from '../../lib/components/DateFilter';
-import { ActionFilter } from './ActionFilter';
-
-
-let colors = ['blue', 'yellow', 'green', 'red', 'purple', 'gray', 'indigo', 'pink', 'orange', 'teal', 'cyan', 'gray-dark'];
-let getColorVar = (variable) => getComputedStyle(document.body).getPropertyValue('--' + variable)
-export class ActionsPie extends Component {
-    constructor(props) {
-        super(props)
-        this.state = {}
-        this.fetchGraph = this.fetchGraph.bind(this);
-        this.fetchGraph()
-    }
-    fetchGraph() {
-        api.get('api/action/trends/?' + toParams(this.props.filters)).then((data) => {
-            data = data.sort((a, b) => b.count - a.count)
-            let color_list = colors.map(color => getColorVar(color));
-            this.setState({
-                data: [{
-                    labels: data.map(item => item.label),
-                    data: data.map(item => item.data && item.data.reduce((prev, d) => prev + d, 0)),
-                    backgroundColor: color_list,
-                    hoverBackgroundColor: color_list,
-                    hoverBorderColor: color_list,
-                    borderColor: color_list,
-                    hoverBorderWidth: 10,
-                    borderWidth: 1
-                }],
-                total: data.reduce((prev, item) => prev + item.count, 0)
-            })
-            this.props.onData && this.props.onData(data)
-        })
-    }
-    componentDidUpdate(prevProps) {
-        if(prevProps.filters !== this.props.filters) this.fetchGraph();
-    }
-    render() {
-        let { data, total } = this.state;
-        return data ? (data[0].labels ? <div style={{position: 'absolute', width: '100%', height: '100%'}}>
-            <h1 style={{position: 'absolute', margin: '0 auto', left: '50%', top: '50%', fontSize: '3rem'}}><div style={{marginLeft: '-50%', marginTop: -30}}>{total}</div></h1>
-            <LineGraph
-            type='doughnut'
-            datasets={data}
-            labels={data[0].labels}
-            />
-        </div>: <p style={{textAlign: 'center', marginTop: '4rem'}}>We couldn't find any matching actions.</p>) : <Loading />;
-
-    }
-}
-
-export class ActionsLineGraph extends Component {
-    constructor(props) {
-        super(props)
-        this.state = {}
-        this.fetchGraph = this.fetchGraph.bind(this);
-        this.fetchGraph()
-    }
-    fetchGraph() {
-        api.get('api/action/trends/?' + toParams(this.props.filters)).then((data) => {
-            data = data.sort((a, b) => b.count - a.count)
-            this.setState({data})
-            this.props.onData && this.props.onData(data)
-        })
-    }
-    componentDidUpdate(prevProps) {
-        if(prevProps.filters !== this.props.filters) this.fetchGraph();
-    }
-    render() {
-        let { data } = this.state;
-        return data ? (data[0].labels ? <LineGraph
-                            datasets={data}
-                            labels={data[0].labels}
-                            /> : <p style={{textAlign: 'center', marginTop: '4rem'}}>We couldn't find any matching actions.</p>) : <Loading />;
-    }
-}
-
-ActionsLineGraph.propTypes = {
-    filters: PropTypes.object.isRequired,
-    onData: PropTypes.func
-}
-
-export class ActionsTable extends Component {
-    constructor(props) {
-        super(props)
-
-        this.state = {
-        }
-        this.fetchGraph = this.fetchGraph.bind(this);
-        this.fetchGraph()
-    }
-    fetchGraph() {
-        api.get('api/action/trends/?' + toParams(this.props.filters)).then((data) => {
-            data = data.sort((a, b) => b.count - a.count)
-            this.setState({data})
-            this.props.onData && this.props.onData(data)
-        })
-    }
-    componentDidUpdate(prevProps) {
-        if(prevProps.filters !== this.props.filters) {
-            this.fetchGraph();
-        }
-    }
-    render() {
-        let { data } = this.state;
-        let { filters } = this.props;
-        return data ? (data[0].labels ? <table className='table'>
-            <tbody>
-                <tr>
-                    <th style={{width: 100}}>Action</th>
-                    {filters.breakdown && <th>Breakdown</th>}
-                    <th style={{width: 50}}>Count</th>
-                </tr>
-                {!filters.breakdown && data.map((item) => <tr key={item.label}>
-                    <td>{item.label}</td>
-                    <td>{item.count}</td>
-                </tr>)}
-                {filters.breakdown && data.filter((item) => item.count > 0).map((item) => [
-                    <tr key={item.label}>
-                        <td rowSpan={item.breakdown.length || 1}>{item.label}</td>
-                        <td className='text-overflow'>{item.breakdown[0] && item.breakdown[0].name}</td>
-                        <td>{item.breakdown[0] && item.breakdown[0].count}</td>
-                    </tr>,
-                    item.breakdown.slice(1).map((i) => <tr key={i.name}>
-                        <td className='text-overflow'>{i.name}</td>
-                        <td>{i.count}</td>
-                    </tr>)
-                ])}
-            </tbody>
-        </table> : <p style={{textAlign: 'center', marginTop: '4rem'}}>We couldn't find any matching actions.</p>) : <Loading />;
-    }
-}
-ActionsTable.propTypes = {
-    filters: PropTypes.object.isRequired,
-    onData: PropTypes.func
-}
-
-class BreakdownFilter extends Component {
-    render() {
-        return <div style={{width: 200, display: 'inline-block'}}>
-            <Select
-                cacheOptions
-                defaultOptions
-                style={{width: 200}}
-                placeholder={"Break down by"}
-                value={this.props.breakdown ? {label: this.props.breakdown, value: this.props.breakdown} : null}
-                onChange={(item) => this.props.onChange(item.value)}
-                styles={selectStyle}
-                options={this.props.properties} />
-        </div>
-    }
-}
-
+import ActionFilter from './ActionFilter';
+import ActionsPie from './ActionsPie'
+import BreakdownFilter from './BreakdownFilter'
+import ActionsTable from './ActionsTable'
 
 export default class ActionsGraph extends Component {
     constructor(props) {
